requirements.txt
python-telegram-bot==13.14
environs==9.5.0
google-cloud-dialogflow==2.16.1
google-cloud-storage==2.5.0